+++ lib/let_me/unauthorized_error.ex
@@ -2,5 +2,9 @@ defmodule LetMe.UnauthorizedError do
   @moduledoc """
   Raised by `c:LetMe.Policy.authorize!/4` if a request is unauthorized.
   """
-  defexception message: "unauthorized"
+  defexception [:message]
+
+  def message(exception) do
+    exception.message || "unauthorized"
+  end
 end